+++ src/utils/utils.cuh
@@ -8,6 +8,6 @@ void* allocGPU(const uint size, const uint count = 1);
 void freeGPU(void* const memory);
 void syncGPU();
 void check_error(const cudaError_t result, const char* const func, const char* const file, const int line);
-void memGpuToCpu(void* const to, const void* const from, const uint size);
-void memCpuToGpu(void* const to, const void* const from, const uint size);
+void memGpuToCpu(void* const to, const void* const from, const size_t size);
+void memCpuToGpu(void* const to, const void* const from, const size_t size);
 void printDevicesInfos();
